chore(LAB-3307): renaming
